update test app to share config form with server, static apps
refactor server for interaction code flow
