+++ test/app/public/index.html
@@ -3,25 +3,16 @@
     <link rel="stylesheet" href="/oidc-app.css"/>
   </head>
   <body class="oidc-app landing">
+    <div class="box">
+      <a class="config-link" href="#" onclick="showConfigForm(event)">⚙️</a>
+      <div class="link selected">Webpack SPA App</div>
+      <a href="/server" onclick="navigateToApp('/server', event)">Server-side Web App</a>
+      <a href="/static" onclick="navigateToApp('/static', event)">Static SPA App</a>
+    </div>
     <div id="root"></div>
     <script src="/oidc-app.js"></script>
     <script type="text/javascript">
       bootstrapLanding();
     </script>
-    <a href="/server">Server-side Login Form</a>
-    <script>
-      function staticApp(e) {
-        e.preventDefault();
-        var config = getAuthJSConfig();
-        var params = {
-          issuer: config.issuer,
-          clientId: config.clientId,
-          redirectUri: config.redirectUri
-        }
-        var url = '/static' + toQueryString(params);
-        window.location.assign(url);
-      }
-    </script>
-    <a href="#" onclick="staticApp(event)">Static App</a>
   </body>
 </html>
