+++ test/app/public/static/index.html
@@ -1,5 +1,9 @@
 <html>
   <head>
+    <link rel="stylesheet" href="/oidc-app.css"/>
+    <!-- for nav functions-->
+    <script src="/oidc-app.js"></script>
+
     <!-- https://global.oktacdn.com/okta-auth-js/3.2.3/okta-auth-js.polyfill.js -->
     <script src="/okta-auth-js.polyfill.js" type="text/javascript"></script>
 
@@ -7,6 +11,12 @@
     <script src="/okta-auth-js.min.js" type="text/javascript"></script>
   </head>
   <body>
+    <div class="box">
+      <a class="config-link" href="#" onclick="showConfigForm(event)">⚙️</a>
+      <a href="/" onclick="navigateToApp('/', event)">Webpack SPA App</a>
+      <a href="/server" onclick="navigateToApp('/server', event)">Server-side Web App</a>
+      <div class="link selected">Static SPA App</div>
+    </div>
     <div id="unauth" style="display: none">
       <b>Please login</b><hr/>
       <a id="login" href="#" onclick="login(event)">Login</a>
