+++ test/app/src/window.ts
@@ -3,8 +3,15 @@
 import * as Cookies from 'js-cookie';
 
 import TestApp from './testApp';
-import { Config, getDefaultConfig, getConfigFromUrl, getConfigFromStorage, clearStorage } from './config';
-import { onSubmitForm, onFormData } from './form';
+import {
+  Config,
+  getDefaultConfig,
+  getConfigFromUrl,
+  getConfigFromStorage,
+  clearStorage,
+  flattenConfig
+} from './config';
+import { onSubmitForm, onFormData, showConfigForm } from './form';
 import { toQueryString } from './util';
 import { FormDataEvent } from './types';
 import { buildWidgetConfig } from './widget';
@@ -20,12 +27,19 @@ declare global {
     getWidgetConfig: () => any;
     getAuthJSConfig: () => any;
     toQueryString: (obj: any) => string;
+    constructAppUrl: (basePath: string) => string;
+    navigateToApp: (appPath: string, event: Event) => void;
+    showConfigForm: (event?: Event) => void;
   }
 }
 
 let app: TestApp;
 let config: Config;
 const rootElem = document.getElementById('root');
+
+window.addEventListener('load', () => {
+  rootElem.classList.add('loaded');
+});
 
 function mount(): TestApp {
   // Create the app as a function of config
@@ -39,32 +53,49 @@ function mount(): TestApp {
   return app;
 }
 
-window.getAuthJSConfig = getDefaultConfig;
-window.toQueryString = toQueryString;
-window.onSubmitForm = onSubmitForm;
-window.onFormData = onFormData;
+Object.assign(window, {
+  toQueryString: toQueryString,
+  onSubmitForm: onSubmitForm,
+  onFormData: onFormData,
 
-// Login page, read config from URL
-window.getWidgetConfig = function(): any {
-  const siwConfig = window.location.search ? getConfigFromUrl() : getDefaultConfig();
-  return buildWidgetConfig(siwConfig);
-};
+  getAuthJSConfig: function(): Config {
+    return window.location.search ? getConfigFromUrl() : getDefaultConfig();
+  },
+  constructAppUrl: function(basePath = ''): string {
+    const config = window.getAuthJSConfig();
+    const queryParams = toQueryString(flattenConfig(config));
+    return basePath + queryParams;
+  },
+  navigateToApp: function(appPath = '', event: Event): void {
+    event.preventDefault();
+    const url = window.constructAppUrl(appPath);
+    window.location.assign(url);
+  },
+  showConfigForm: function(event?: Event) {
+    if (event) {
+      event.preventDefault();
+    }
+    const config = window.getAuthJSConfig();
+    showConfigForm(config);
+  },
+  // Login page, read config from URL
+  getWidgetConfig: function(): any {
+    const siwConfig = window.location.search ? getConfigFromUrl() : getDefaultConfig();
+    return buildWidgetConfig(siwConfig);
+  },
 
-// Regular landing, read config from URL
-window.bootstrapLanding = function(): void {
-  config = window.location.search ? getConfigFromUrl() : getDefaultConfig();
-  mount();
-  app.bootstrapHome();
-};
+  // Regular landing, read config from URL
+  bootstrapLanding: function(): void {
+    config = window.location.search ? getConfigFromUrl() : getDefaultConfig();
+    mount();
+    app.bootstrapHome();
+  },
 
-// Callback, read config from storage
-window.bootstrapCallback = function(): void {
-  config = getConfigFromStorage() || getDefaultConfig();
-  clearStorage();
-  mount();
-  app.bootstrapCallback();
-};
-
-window.addEventListener('load', () => {
-  rootElem.classList.add('loaded');
+  // Callback, read config from storage
+  bootstrapCallback: function(): void {
+    config = getConfigFromStorage() || getDefaultConfig();
+    clearStorage();
+    mount();
+    app.bootstrapCallback();
+  }
 });
